+++ client/src/redux/actions/userActions.js
@@ -1,6 +1,7 @@
 import { ActionTypes } from "../constants/actionTypes";
 import axios from "axios";
 
+/**** Create ****/
 /* 유저 생성 */
 export const createUserData = (inputs) => {
 	return async (dispatch, getState) => {
@@ -8,16 +9,17 @@ export const createUserData = (inputs) => {
 		try {
 			// 메타마스크 reducer에서 주소 가져옴
 			let accounts = getState().metamask.accounts;
+			if (accounts.length>0) {
+				const url = "http://localhost:5000/users/";
+				await axios.post(url, { ...inputs, address: accounts[0] });
 
-			const url = "http://localhost:5000/users/";
-			const createData = (
-				await axios.post(url, { ...inputs, address: accounts[0] })
-			).data;
-			
-			dispatch({
-				type: ActionTypes.USER_CREATE_SUCCESS,
-				payload: createData,
-			});
+				dispatch({ type: ActionTypes.USER_CREATE_SUCCESS });
+			} else {
+				dispatch({
+					type: ActionTypes.USER_DATA_FAIL,
+					payload: "Account is not found",
+				});
+			}
 		} catch (error) {
 			dispatch({
 				type: ActionTypes.USER_DATA_FAIL,
@@ -27,6 +29,7 @@ export const createUserData = (inputs) => {
 	};
 };
 
+/**** Read ****/
 /* 유저정보 불러오기 */
 export const readUserData = () => {
 	return async (dispatch, getState) => {
@@ -34,14 +37,20 @@ export const readUserData = () => {
 		try {
 			// 메타마스크 reducer에서 주소 가져옴
 			let accounts = getState().metamask.accounts;
-
-			const url = `http://localhost:5000/users/${accounts[0]}`;
-			const userInfo = (await axios.get(url)).data;
-
-			dispatch({
-				type: ActionTypes.USER_READ_SUCCESS,
-				payload: userInfo,
-			});
+			if (accounts.length>0) {
+				const url = `http://localhost:5000/users/${accounts[0]}`;
+				const userInfo = (await axios.get(url)).data;
+	
+				dispatch({
+					type: ActionTypes.USER_READ_SUCCESS,
+					payload: userInfo,
+				});
+			} else {
+				dispatch({
+					type: ActionTypes.USER_DATA_FAIL,
+					payload: "Account is not found",
+				});
+			}
 		} catch (error) {
 			dispatch({
 				type: ActionTypes.USER_DATA_FAIL,
@@ -51,6 +60,7 @@ export const readUserData = () => {
 	};
 };
 
+/**** Update ****/
 /* 유저정보 수정 */
 export const updateUserData = (inputs) => {
 	return async (dispatch, getState) => {
@@ -58,15 +68,21 @@ export const updateUserData = (inputs) => {
 		try {
 			// 메타마스크 reducer에서 주소 가져옴
 			let accounts = getState().metamask.accounts;
-
-			const url = `http://localhost:5000/users/${accounts[0]}`;
-
-			await axios.patch(url, inputs);
-
-			dispatch({
-				type: ActionTypes.USER_UPDATE_SUCCESS,
-				payload: inputs,
-			});
+			if (accounts.length>0) {
+				const url = `http://localhost:5000/users/${accounts[0]}`;
+	
+				await axios.patch(url, inputs);
+	
+				dispatch({
+					type: ActionTypes.USER_UPDATE_SUCCESS,
+					payload: inputs,
+				});
+			} else {
+				dispatch({
+					type: ActionTypes.USER_DATA_FAIL,
+					payload: "Account is not found",
+				});
+			}
 		} catch (error) {
 			dispatch({
 				type: ActionTypes.USER_DATA_FAIL,
@@ -76,6 +92,7 @@ export const updateUserData = (inputs) => {
 	};
 };
 
+/**** Delete ****/
 /* 유저 삭제 */
 export const deleteUserData = () => {
 	return async (dispatch, getState) => {
@@ -83,13 +100,19 @@ export const deleteUserData = () => {
 		try {
 			// 메타마스크 reducer에서 주소 가져옴
 			let accounts = getState().metamask.accounts;
-
-			const url = `http://localhost:5000/users/${accounts[0]}`;
-			const result = (await axios.delete(url)).data;
-
-			dispatch({
-				type: ActionTypes.USER_DELETE_SUCCESS,
-			});
+			if (accounts.length>0) {
+				const url = `http://localhost:5000/users/${accounts[0]}`;
+				const result = (await axios.delete(url)).data;
+	
+				dispatch({
+					type: ActionTypes.USER_DELETE_SUCCESS,
+				});
+			} else {
+				dispatch({
+					type: ActionTypes.USER_DATA_FAIL,
+					payload: "Account is not found",
+				});
+			}
 		} catch (error) {
 			dispatch({
 				type: ActionTypes.USER_DATA_FAIL,
